requirements.txt.txt
streamlit
yfinance
pandas
plotly
google-generativeai
streamlit-searchbox